Refactor command handling and move commands into Terminal.
Split out the parameter handling logic of commands from the actual command,
this allows us to reference CSI like commands from unit tests and makes
porting the esctest test suite easier.

Added CBT tests.
